admin 예외처리 추가
* basic setting

* add admin api

* update member api

* update admin api

* admin 예외처리 추가

+++ src/main/java/landvibe/test/service/AdminService.java
@@ -1,12 +1,14 @@
 package landvibe.test.service;
 
 import landvibe.test.domain.Member;
+import landvibe.test.exception.RuralException;
 import landvibe.test.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -14,18 +16,30 @@ public class AdminService {
     private final MemberRepository memberRepository;
 
     public List<Member> findMembersApproveFalse() {
-        List<Member> members = memberRepository.findByApproveFalse();
-        return members;
+        Optional<List<Member>> byApproveFalse = memberRepository.findByApproveFalse();
+        if(byApproveFalse.isEmpty()){
+            return null;
+        }
+        return byApproveFalse.get();
     }
 
     @Transactional
-    public void approveMember(Long memberId) {
-        Member member = memberRepository.findById(memberId).get();
-        member.setApprove(true);
+    public void approveMember(Long memberId) throws RuralException {
+        Optional<Member> byId = memberRepository.findById(memberId);
+        if(byId.isEmpty()){
+            // 존재하지 않는 회원
+            throw new RuralException("존재하지 않는 회원");
+        }
+        byId.get().setApprove(true);
     }
 
     @Transactional
-    public void refuseMember(Long memberId) {
+    public void refuseMember(Long memberId) throws RuralException {
+        Optional<Member> byId = memberRepository.findById(memberId);
+        if(byId.isEmpty()){
+            // 존재하지 않는 회원
+            throw new RuralException("존재하지 않는 회원");
+        }
         memberRepository.deleteById(memberId);
     }
 }
